cleaned up repo and separated necessary files
