--- app/api/operations/get_prediction.py
@@ -1,6 +1,0 @@
-import uuid
-from dataclasses import dataclass
-from typing import Any
-
-from fastapi import APIRouter
-
