+++ scripts/launch_api.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+echo "hello"
+
+cd "$(dirname "$0")/.." || exit
+
+# uvicorn app.main:app --reload
+# python -m main
